pyproject.toml
[project]
name = "dod-v2"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "langchain-core>=0.3.79",
    "langchain-openai>=0.3.35",
    "langgraph>=0.6.10",
    "openai>=1.0.0",
    "pandas>=2.0.0",
    "python-dotenv>=1.0.0",
]
